Refactor import paths and implement activity, theory, and visualization components for various list types
- Updated import paths from "../tipos/" to "../types/" in list-activity.tsx, list-theory.tsx, and list-visualization.tsx.
- Created LcActivity, LcTheory, and LcVisualization components for Circular List.
- Created LddeActivity, LddeTheory, and LddeVisualization components for Doubly Linked List.
- Implemented QuestaoPage component for LDSE activities with session management and activity handling.
- Added LdseTheory and LdseVisualization components for Doubly Linked List theory and visualization.
- Created LeeActivity, LeeTheory, and LeeVisualization components for Static Linked List.
- Developed LesActivity, LesTheory, and LesVisualization components for Sequential Static List.
