+++ src/app/estruturas/lista/types/lc/activity.tsx
