+++ src/app/estruturas/lista/types/lc/theory.tsx
