+++ src/app/estruturas/lista/types/lc/visualization.tsx
