+++ src/app/estruturas/lista/types/ldde/activity.tsx
