+++ src/app/estruturas/lista/types/ldde/theory.tsx
