+++ src/app/estruturas/lista/types/ldde/visualization.tsx
