+++ src/app/estruturas/lista/types/ldse/activity.tsx
