+++ src/app/estruturas/lista/types/ldse/theory.tsx
