+++ src/app/estruturas/lista/types/lee/activity.tsx
