+++ src/app/estruturas/lista/types/lee/theory.tsx
